feat(www): rename /create to /studio and add navbar CTA on home
