+++ CLAUDE.md
@@ -1,6 +1,6 @@
 # dotUI
 
-This is the dotUI repository — a design-system builder. Users compose a complete design system at dotui.org/create — colors, typography, icons, density, radius, per-component styles — preview every change live on real components, and export it as code they own: into their codebase via the shadcn CLI, or straight into v0 — with Bolt, Lovable, Figma and more planned. It's built on React Aria Components with Tailwind CSS v4 and tailwind-variants, themed by an OKLCH color engine, and distributed as a shadcn-compatible registry served from a TanStack Start app.
+This is the dotUI repository — a design-system builder. Users compose a complete design system at dotui.org/studio — colors, typography, icons, density, radius, per-component styles — preview every change live on real components, and export it as code they own: into their codebase via the shadcn CLI, or straight into v0 — with Bolt, Lovable, Figma and more planned. It's built on React Aria Components with Tailwind CSS v4 and tailwind-variants, themed by an OKLCH color engine, and distributed as a shadcn-compatible registry served from a TanStack Start app.
 
 ## Product direction
 
@@ -33,13 +33,13 @@ An honest snapshot (August 2026) to calibrate against. Task-level work is tracke
 
 - **Solid:** the docs (one last revision pass pending) and the components — they look and behave well individually. Cross-component composition is less proven: treat "these compose cleanly" as a claim to verify, not an assumption.
 - **Known-weak:** the components page (several demos misbehave); the presets page (very slow); the presets themselves on both landing and presets page — due to be replaced by fewer, high-fidelity presets of popular systems (Linear, Vercel, …); chart colors (need a full rewrite); site performance overall (never audited — a docs site must feel instant).
-- **In progress — the /create panel rewrite** (tracking: issue #666). The panel-lab drill-in panel (`www/src/modules/panel-lab/`) is /create's real panel; its chrome (switcher, reset, search, save, export) is wired to the create engine, and chapters are being wired one at a time — one short-lived PR per chapter, merged to main. Rules: the panel schema (`panel-lab/state.ts`) is the source of truth — wiring a chapter means making the engine/preview/export respond to its axes (reuse engine machinery where an axis matches, change the engine where it diverges, build what's new; never bend the panel to old engine params). A chapter carries a WIP chip until its values actually drive the preview and export. The old panel (`www/src/modules/create/panel/`, unmounted) holds working wiring logic — mine it per chapter, delete what's absorbed in the same PR.
+- **In progress — the /studio panel rewrite** (tracking: issue #666). The panel-lab drill-in panel (`www/src/modules/panel-lab/`) is /studio's real panel; its chrome (switcher, reset, search, save, export) is wired to the create engine, and chapters are being wired one at a time — one short-lived PR per chapter, merged to main. Rules: the panel schema (`panel-lab/state.ts`) is the source of truth — wiring a chapter means making the engine/preview/export respond to its axes (reuse engine machinery where an axis matches, change the engine where it diverges, build what's new; never bend the panel to old engine params). A chapter carries a WIP chip until its values actually drive the preview and export. The old panel (`www/src/modules/create/panel/`, unmounted) holds working wiring logic — mine it per chapter, delete what's absorbed in the same PR.
 - **Queued for rewrite — don't deepen:** the color system gets a complete rewrite — the goal is really good foundations, not incremental fixes; it starts with a playground to verify generated colors visually (current output is average). The publisher is also queued (see below), and the registry should get a small refactor and clean up. In these areas, verify behavior by reading the code and keep changes shallow.
 - **AI slop:** much of the codebase was written by AI agents and hasn't all been walked through. When working in an area, leave it simpler than you found it — cut needless comments, indirection, and dead options rather than matching the existing style.
 
 ## Structure
 
-- `www/` — the dotui.org app: docs, the /create builder, and the registry endpoints (TanStack Start + Vite, fumadocs-mdx, Tailwind v4, tailwind-variants). Registry source lives in `www/src/registry/` — see Registry below.
+- `www/` — the dotui.org app: docs, the /studio builder, and the registry endpoints (TanStack Start + Vite, fumadocs-mdx, Tailwind v4, tailwind-variants). Registry source lives in `www/src/registry/` — see Registry below.
 - `packages/colors` — `@dotui/colors`, the OKLCH color engine (private, consumed by www).
 - Starter themes and the Tailwind plugins (`tailwindcss-autocontrast`, `tailwindcss-with`) live in standalone repos, consumed from npm — their source is not here.
