+++ www/content/docs/(root)/index.mdx
@@ -3,7 +3,7 @@ title: Introduction
 description: A design-system builder and component registry. You compose it, preview it live, and own the code.
 ---
 
-**dotUI is a design-system builder.** You compose your system at [dotui.org/create](https://dotui.org/create) — colors, typography, icons, density, radius, per-component styles — preview every change on live components, and export code you own: into your codebase via the shadcn CLI, or straight into [v0](https://v0.dev), with more targets planned.
+**dotUI is a design-system builder.** You compose your system at [dotui.org/studio](https://dotui.org/studio) — colors, typography, icons, density, radius, per-component styles — preview every change on live components, and export code you own: into your codebase via the shadcn CLI, or straight into [v0](https://v0.dev), with more targets planned.
 
 Open code alone doesn't get you there: you own the files, but the decisions inside them are still someone else's. dotUI makes the decisions yours. The output is not our library with your colors — it is your design system.
