+++ www/content/docs/(root)/installation.mdx
@@ -33,10 +33,10 @@ This writes the `@dotui` registry into `components.json`:
 }
 ```
 
-The empty `?preset=` is a placeholder — a style exported from the [editor](/create) fills it in so every component installs pre-themed.
+The empty `?preset=` is a placeholder — a style exported from the [editor](/studio) fills it in so every component installs pre-themed.
 
 <Alert>
-  Building a style in the [editor](/create)? Its export dialog gives you an
+  Building a style in the [editor](/studio)? Its export dialog gives you an
   `init` command with your theme baked in, so every component installs
   pre-styled to your system.
 </Alert>
