+++ www/content/docs/components/chart.mdx
@@ -86,7 +86,7 @@ Because config is separate from data, a key with no series of its own — an axi
 
 ## Theming
 
-Series colors come from a categorical palette of five slots, `--chart-1` through `--chart-5`. Your design system derives them from its theme hues, so charts stay on-brand and adapt to light and dark automatically — no per-mode values to maintain. The palette is editable under **Chart colors** in the [editor](/create), and updates live as you tweak it on the [charts page](/docs/charts).
+Series colors come from a categorical palette of five slots, `--chart-1` through `--chart-5`. Your design system derives them from its theme hues, so charts stay on-brand and adapt to light and dark automatically — no per-mode values to maintain. The palette is editable under **Chart colors** in the [editor](/studio), and updates live as you tweak it on the [charts page](/docs/charts).
 
 Point each series at a slot through its config color:
